Fix test failure from report function

+++ utils/report.ts
@@ -1,8 +1,12 @@
 // Import
-import { writeFileSync } from 'fs';
+import { existsSync, writeFileSync } from 'fs';
 
 // Report function
 const report = (routine: string, index: number, time: number): void => {
+  if (!existsSync('./tests/reports')) {
+    return;
+  }
+
   writeFileSync(
     './tests/reports/report.txt',
     `Routine ${routine} - case ${index}: ${time}ms\n`,
